extract common code
