+++ examples/grpc/python/simple_ecu/ecu.py
@@ -3,13 +3,13 @@
 from __future__ import print_function
 
 import os
-import random
 import time
+import binascii
 
 import grpc
 
 import sys
-sys.path.append('generated')
+sys.path.append('../common/generated')
 
 import network_api_pb2
 import network_api_pb2_grpc
@@ -17,58 +17,11 @@ import system_api_pb2
 import system_api_pb2_grpc
 import common_pb2
 
+sys.path.append('../common')
+import helper
+from helper import *
+
 from threading import Thread, Timer
-##################### START BOILERPLATE ####################################################
-
-import hashlib
-import posixpath
-import ntpath
-import itertools
-
-def get_sha256(file):
-        f = open(file,"rb")
-        bytes = f.read() # read entire file as bytes
-        readable_hash = hashlib.sha256(bytes).hexdigest();
-        return readable_hash
-
-def generate_data(file, dest_path, chunk_size, sha256):
-    for x in itertools.count(start=0):
-        if x == 0:
-                fileDescription = system_api_pb2.FileDescription(sha256 = sha256, path = dest_path)
-                yield system_api_pb2.FileUploadRequest(fileDescription = fileDescription)
-        else:
-                buf = file.read(chunk_size)
-                if not buf:
-                        break
-                yield system_api_pb2.FileUploadRequest(chunk = buf)   
-
-def upload_file(stub, path, dest_path):
-     sha256 = get_sha256(path)
-     print(sha256)
-     file = open(path, "rb")  
-
-     # make sure path is unix style (necessary for windows, and does no harm om linux)
-     upload_iterator = generate_data(file, dest_path.replace(ntpath.sep, posixpath.sep), 1000000, sha256)
-     response = stub.UploadFile(upload_iterator)
-     print("uploaded", path, response)
-
-from glob import glob
-
-def upload_folder(system_stub, folder):
-     files = [y for x in os.walk(folder) for y in glob(os.path.join(x[0], '*')) if not os.path.isdir(y)]
-     for file in files:
-            upload_file(system_stub, file, file.replace(folder, ""))
-
-def reload_configuration(system_stub):
-      request = common_pb2.Empty()
-      response = system_stub.ReloadConfiguration(request, timeout=60000)
-      print(response)
-
-def check_license(system_stub):
-    status = system_stub.GetLicenseInfo(common_pb2.Empty()).status
-    assert status == system_api_pb2.LicenseStatus.VALID, "Check your license, status is: %d" % status
-
-##################### END BOILERPLATE ####################################################
 
 def read_signal(stub, signal):
     read_info = network_api_pb2.SignalIds(signalId=[signal])
